+++ CHANGELOG.md
@@ -16,9 +16,13 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Home page:
         - Own avatar displaying with question marks when signing in or signing up. ([#978], [#950])
+    - Chat page:
+        - Tapping on image opens gallery when being in context menu. ([#983], [#979])
 
 [#950]: /../../issues/950
 [#978]: /../../pull/978
+[#979]: /../../issues/979
+[#983]: /../../pull/978